Moved the useStreamgageData hook to a common file, condensed its parameters
