+++ app/client/src/components/shared/LocationMap.js
@@ -32,8 +32,6 @@ import {
   getUniqueWaterbodies,
 } from 'utils/mapFunctions';
 import MapErrorBoundary from 'components/shared/ErrorBoundary.MapErrorBoundary';
-// config
-import { usgsStaParameters } from 'config/usgsStaParameters';
 // contexts
 import { useFetchedDataDispatch, useFetchedDataState } from 'contexts/FetchedData';
 import { LocationSearchContext } from 'contexts/locationSearch';
@@ -57,6 +55,7 @@ import {
   useDynamicPopup,
   useGeometryUtils,
   useSharedLayers,
+  useStreamgageData,
   useWaterbodyHighlight,
   useWaterbodyFeatures,
 } from 'utils/hooks';
@@ -79,79 +78,6 @@ import { colors } from 'styles/index.js';
 // turns an array into a string for the service queries
 function createQueryString(array) {
   return `'${array.join("', '")}'`;
-}
-
-function useStreamgageData(fetchedData, fetchStatus) {
-  const [normalizedUsgsStreamgages, setNormalizedUsgsStreamgages] = useState(
-    [],
-  );
-
-  // normalize USGS streamgages data with monitoring stations data
-  useEffect(() => {
-    if (fetchStatus !== 'success' || !fetchedData.value) return;
-
-    const gages = fetchedData.value.map((gage) => {
-      const streamgageMeasurements = { primary: [], secondary: [] };
-
-      [...gage.Datastreams]
-        .filter((item) => item.Observations.length > 0)
-        .forEach((item) => {
-          const observation = item.Observations[0];
-          const parameterCode = item.properties.ParameterCode;
-          const parameterDesc = item.description.split(' / USGS-')[0];
-          const parameterUnit = item.unitOfMeasurement;
-
-          let measurement = observation.result;
-          // convert measurements recorded in celsius to fahrenheit
-          if (['00010', '00020', '85583'].includes(parameterCode)) {
-            measurement = measurement * (9 / 5) + 32;
-          }
-
-          const matchedParam = usgsStaParameters.find((p) => {
-            return p.staParameterCode === parameterCode;
-          });
-
-          const data = {
-            parameterCategory: matchedParam?.hmwCategory || 'exclude',
-            parameterOrder: matchedParam?.hmwOrder || 0,
-            parameterName: matchedParam?.hmwName || parameterDesc,
-            parameterUsgsName: matchedParam?.staDescription || parameterDesc,
-            parameterCode,
-            measurement,
-            datetime: new Date(observation.phenomenonTime).toLocaleString(),
-            dailyAverages: [],
-            unitAbbr: matchedParam?.hmwUnits || parameterUnit.symbol,
-            unitName: parameterUnit.name,
-          };
-
-          if (data.parameterCategory === 'primary') {
-            streamgageMeasurements.primary.push(data);
-          }
-
-          if (data.parameterCategory === 'secondary') {
-            streamgageMeasurements.secondary.push(data);
-          }
-        });
-
-      return {
-        monitoringType: 'Current Water Conditions',
-        siteId: gage.properties.monitoringLocationNumber,
-        orgId: gage.properties.agencyCode,
-        orgName: gage.properties.agency,
-        locationLongitude: gage.Locations[0].location.coordinates[0],
-        locationLatitude: gage.Locations[0].location.coordinates[1],
-        locationName: gage.properties.monitoringLocationName,
-        locationType: gage.properties.monitoringLocationType,
-        locationUrl: gage.properties.monitoringLocationUrl,
-        // usgs streamgage specific properties:
-        streamgageMeasurements,
-      };
-    });
-
-    setNormalizedUsgsStreamgages(gages);
-  }, [fetchedData, fetchStatus]);
-
-  return normalizedUsgsStreamgages;
 }
 
 const mapPadding = 20;
@@ -1308,7 +1234,7 @@ function LocationMap({ layout = 'narrow', windowHeight, children }: Props) {
     [services, fetchedDataDispatch],
   );
 
-  const normalizedUsgsStreamgages = useStreamgageData(usgsStreamgages.data, usgsStreamgages.status);
+  const normalizedUsgsStreamgages = useStreamgageData(usgsStreamgages);
 
   useEffect(() => {
     if (!usgsStreamgagesLayer || !normalizedUsgsStreamgages.length) return;
